requirements.txt
requests==2.22.0
robotframework==3.1.2
docker==4.1.0
RESTinstance==1.0.2